fix: issue #3: duplicate code block
