+++ package.json
@@ -31,18 +31,35 @@
     "@radix-ui/colors": "^1.0.1",
     "@radix-ui/react-icons": "^1.3.0",
     "@tiptap/core": "^2.0.3",
+    "@tiptap/extension-blockquote": "^2.0.4",
+    "@tiptap/extension-bold": "^2.0.4",
+    "@tiptap/extension-bullet-list": "^2.0.4",
+    "@tiptap/extension-code": "^2.0.4",
     "@tiptap/extension-code-block": "^2.0.3",
     "@tiptap/extension-code-block-lowlight": "^2.0.3",
+    "@tiptap/extension-document": "^2.0.4",
+    "@tiptap/extension-dropcursor": "^2.0.4",
+    "@tiptap/extension-gapcursor": "^2.0.4",
+    "@tiptap/extension-hard-break": "^2.0.4",
+    "@tiptap/extension-heading": "^2.0.4",
     "@tiptap/extension-highlight": "^2.0.3",
+    "@tiptap/extension-history": "^2.0.4",
+    "@tiptap/extension-horizontal-rule": "^2.0.4",
     "@tiptap/extension-image": "^2.0.3",
+    "@tiptap/extension-italic": "^2.0.4",
     "@tiptap/extension-link": "^2.0.3",
+    "@tiptap/extension-list-item": "^2.0.4",
+    "@tiptap/extension-ordered-list": "^2.0.4",
+    "@tiptap/extension-paragraph": "^2.0.4",
     "@tiptap/extension-placeholder": "^2.0.4",
+    "@tiptap/extension-strike": "^2.0.4",
     "@tiptap/extension-table": "^2.0.3",
     "@tiptap/extension-table-cell": "^2.0.3",
     "@tiptap/extension-table-header": "^2.0.3",
     "@tiptap/extension-table-row": "^2.0.3",
     "@tiptap/extension-task-item": "^2.0.3",
     "@tiptap/extension-task-list": "^2.0.3",
+    "@tiptap/extension-text": "^2.0.4",
     "@tiptap/extension-typography": "^2.0.3",
     "@tiptap/pm": "^2.0.3",
     "@tiptap/react": "^2.0.3",
